feat: migrate browsing to usecases
