+++ tests/modules/reporter_test.go
@@ -29,7 +29,7 @@ import (
 	"github.com/sergeii/swat4master/internal/testutils/factories"
 )
 
-func makeApp(extra ...fx.Option) (*fx.App, func()) {
+func makeAppWithReporter(extra ...fx.Option) (*fx.App, func()) {
 	fxopts := []fx.Option{
 		application.Module,
 		fx.Provide(func() config.Config {
@@ -51,7 +51,7 @@ func makeApp(extra ...fx.Option) (*fx.App, func()) {
 
 func TestReporter_Available_OK(t *testing.T) {
 	ctx := context.TODO()
-	app, cancel := makeApp()
+	app, cancel := makeAppWithReporter()
 	defer cancel()
 	app.Start(ctx) // nolint: errcheck
 
@@ -62,7 +62,7 @@ func TestReporter_Available_OK(t *testing.T) {
 func TestReporter_Challenge_OK(t *testing.T) {
 	ctx := context.TODO()
 
-	app, cancel := makeApp()
+	app, cancel := makeAppWithReporter()
 	defer cancel()
 	app.Start(ctx) // nolint: errcheck
 
@@ -101,7 +101,7 @@ func TestReporter_Challenge_InvalidInstanceID(t *testing.T) {
 	for _, tt := range tests {
 		t.Run(tt.name, func(t *testing.T) {
 			ctx := context.TODO()
-			app, cancel := makeApp()
+			app, cancel := makeAppWithReporter()
 			defer cancel()
 			app.Start(ctx) // nolint: errcheck
 
@@ -125,7 +125,7 @@ func TestReporter_Heartbeat_OK(t *testing.T) {
 	var probeRepo repositories.ProbeRepository
 
 	ctx := context.TODO()
-	app, cancel := makeApp(
+	app, cancel := makeAppWithReporter(
 		fx.Populate(&serverRepo, &instanceRepo, &probeRepo),
 	)
 	defer cancel()
@@ -155,7 +155,7 @@ func TestReporter_Heartbeat_ServerIsAddedAndThenUpdated(t *testing.T) {
 	var probeRepo repositories.ProbeRepository
 
 	ctx := context.TODO()
-	app, cancel := makeApp(
+	app, cancel := makeAppWithReporter(
 		fx.Populate(&serverRepo, &instanceRepo, &probeRepo),
 	)
 	defer cancel()
@@ -218,7 +218,7 @@ func TestReporter_Heartbeat_ServerIsUpdatedWithNewInstanceID(t *testing.T) {
 	var probeRepo repositories.ProbeRepository
 
 	ctx := context.TODO()
-	app, cancel := makeApp(
+	app, cancel := makeAppWithReporter(
 		fx.Populate(&serverRepo, &instanceRepo, &probeRepo),
 	)
 	defer cancel()
@@ -319,7 +319,7 @@ func TestReporter_Heartbeat_ServerPortIsDiscovered(t *testing.T) {
 			var probeRepo repositories.ProbeRepository
 
 			ctx := context.TODO()
-			app, cancel := makeApp(
+			app, cancel := makeAppWithReporter(
 				fx.Populate(&serverRepo, &instanceRepo, &probeRepo),
 			)
 			defer cancel()
@@ -370,7 +370,7 @@ func TestReporter_Heartbeat_ServerIsRefreshed(t *testing.T) {
 	var serverRepo repositories.ServerRepository
 
 	ctx := context.TODO()
-	app, cancel := makeApp(
+	app, cancel := makeAppWithReporter(
 		fx.Populate(&serverRepo),
 	)
 	defer cancel()
@@ -403,7 +403,7 @@ func TestReporter_Heartbeat_ServerIsRemoved(t *testing.T) {
 	var serverRepo repositories.ServerRepository
 
 	ctx := context.TODO()
-	app, cancel := makeApp(
+	app, cancel := makeAppWithReporter(
 		fx.Populate(&serverRepo),
 	)
 	defer cancel()
@@ -491,7 +491,7 @@ func TestReporter_Heartbeat_ServerRemovalIsValidated(t *testing.T) {
 			var metrics *monitoring.MetricService
 
 			ctx := context.TODO()
-			app, cancel := makeApp(
+			app, cancel := makeAppWithReporter(
 				fx.Populate(&serverRepo, &instanceRepo, &metrics),
 			)
 			defer cancel()
@@ -583,7 +583,7 @@ func TestReporter_Heartbeat_InvalidPayload(t *testing.T) {
 	for _, tt := range tests {
 		t.Run(tt.name, func(t *testing.T) {
 			ctx := context.TODO()
-			app, cancel := makeApp()
+			app, cancel := makeAppWithReporter()
 			defer cancel()
 			app.Start(ctx) // nolint: errcheck
 
@@ -605,7 +605,7 @@ func TestReporter_Keepalive_ServerIsRefreshed(t *testing.T) {
 	var serverRepo repositories.ServerRepository
 
 	ctx := context.TODO()
-	app, cancel := makeApp(
+	app, cancel := makeAppWithReporter(
 		fx.Populate(&serverRepo),
 	)
 	defer cancel()
@@ -698,7 +698,7 @@ func TestReporter_Keepalive_Errors(t *testing.T) {
 			var instanceRepo repositories.InstanceRepository
 
 			ctx := context.TODO()
-			app, cancel := makeApp(
+			app, cancel := makeAppWithReporter(
 				fx.Populate(&serverRepo, &instanceRepo),
 			)
 			defer cancel()
